+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 Yii1 Dependency Injection extension
 ===================================
 
+1.0.2 Under Development
+-----------------------
+
+- Bug: Fixed `ResolvesComponentViaDI` triggers PHP Error on attempt to get unexisting component (klimov-paul)
+- Bug: Fixed `ResolvesComponentViaDI` triggers PHP Error on attempt to get component defined by string class name instead of array (klimov-paul)
+
+
 1.0.1, August 3, 2023
 ---------------------